Added tests and then methods in the account class to pull data from the Account Repository

+++ src/main/java/com/qa/persistence/repository/AccountRepository.java
@@ -4,8 +4,7 @@ public interface AccountRepository {
 
 	String getAllAccounts();
 	String createAccount(String account);
-	String deleteAccount(Long id);
-	String updateAccount(Long id, String account);
+	String deleteAccount(Integer id);
+	String updateAccount(Integer id, String account);
 
-	
 }
